Gatekeeper minor changes, proxy part X
Made sure is forwarded to proxy correctly, might have tiny things to adjust related to security groups.
Tested without custom sg for th and it works, so the sg is causing forwarding issues.
Also changed the proxy function so that it works for SQL queries (minor changes), need to test to see if that works with cluster